Add plot_funcs.py module
 - Plot functions from run_eda.py moved into the new module
